RFC Update the README re: how to install locally
Add instructions for finding the Cura configration folder.

Add potentially correct instructions for how to copy your plugin
to the plugins directory for testing.  The existing instructions
re: zipping the directory with a .plugin and dragging and dropping
it did not work for me;  nevertheless, I kept those instructions
in here in case that's how you have to package for the Marketplace.

If someone knows what the correct instructions are here, I'd be
happy to update as appropriate.